Fix ESP32 compilation: Platform-specific type compatibility
BREAKTHROUGH: Successfully compiled BasicInterpreter for ESP32-S3!
- Program storage: 866KB (66% of 1.31MB)
- Dynamic memory: 24KB (7% of 327KB)

ESP32 Compatibility Fixes:
1. Fixed nested C comment in ArduinoASTInterpreter.h line 11
2. Removed CommandValue from public API (internal type)
3. Changed all std::variant<int> → std::variant<long int> operations
   - ArduinoLibraryRegistry.cpp: convertToInt(), capacitiveSensor methods
   - ASTInterpreter.cpp: All holds_alternative<int> and get<int> calls
4. Fixed FlexibleCommandValue visitor lambda to check int32_t not int
5. Fixed enum counter assignment to cast int → long int
6. Fixed OUTPUT_STREAM << on ESP32 (Serial.println instead of <<)

Root Cause: ESP32 uses long int (32-bit) in variant types while int is
a distinct type from the compiler's perspective. Template instantiation
requires exact type matches for std::variant operations.

This validates v20.0.0 RTTI-free architecture for ESP32 deployment!

+++ src/cpp/ArduinoLibraryRegistry.cpp
@@ -16,8 +16,8 @@ namespace arduino_interpreter {
 
 // Helper conversion functions
 static int32_t convertToInt(const CommandValue& value) {
-    if (std::holds_alternative<int>(value)) {
-        return std::get<int>(value);
+    if (std::holds_alternative<long int>(value)) {
+        return std::get<long int>(value);
     } else if (std::holds_alternative<double>(value)) {
         return static_cast<int32_t>(std::get<double>(value));
     } else if (std::holds_alternative<bool>(value)) {
@@ -240,8 +240,8 @@ void ArduinoLibraryRegistry::registerCapacitiveSensorLibrary() {
     capSensor.internalMethods["capacitiveSensor"] = [](const std::vector<CommandValue>& args, ASTInterpreter* interpreter) -> CommandValue {
         // Get sample count argument (default 30)
         int32_t sampleCount = 30;
-        if (!args.empty() && std::holds_alternative<int>(args[0])) {
-            sampleCount = std::get<int>(args[0]);
+        if (!args.empty() && std::holds_alternative<long int>(args[0])) {
+            sampleCount = std::get<long int>(args[0]);
         }
 
         // Get external value from parent app provider
@@ -257,8 +257,8 @@ void ArduinoLibraryRegistry::registerCapacitiveSensorLibrary() {
     capSensor.internalMethods["capacitiveSensorRaw"] = [](const std::vector<CommandValue>& args, ASTInterpreter* interpreter) -> CommandValue {
         // Get sample count argument (default 30)
         int32_t sampleCount = 30;
-        if (!args.empty() && std::holds_alternative<int>(args[0])) {
-            sampleCount = std::get<int>(args[0]);
+        if (!args.empty() && std::holds_alternative<long int>(args[0])) {
+            sampleCount = std::get<long int>(args[0]);
         }
 
         // Get external value from parent app provider
